Improve unit detection
While this approach is more involved and requires handling a few more
special cases, it leads to a much nicer display in the front end because
we can show the size of each package instead of the standard units which
can look odd. We can still fall back on those if we can't figure out a
regular size.