Fix missing history
Two issues: First, when generating history from scratch, it would
overwrite the history for one store. Second, the Github action would not
correctly download the latest history so it would also not fill in
history but rather start fresh.

This fixes both my making it an error if history is missing and by
ensuring it's correctly filled in when generating from scratch.